+++ lib/internals/src/public_api.ts
@@ -9,3 +9,4 @@ export { ensureStateMetadata } from './utils/ensure-state-metadata';
 export { ensureRepository } from './utils/ensure-repository';
 export { ngxsDeepFreeze } from './utils/freeze';
 export { getRepository } from './utils/get-repository';
+export { NgxsDataInjector } from './services/ngxs-data-injector.service';