+++ lib/storage/src/ngxs-data-storage-plugin.service.ts
@@ -1,5 +1,5 @@
 import { isPlatformServer } from '@angular/common';
-import { Inject, Injectable, Injector, PLATFORM_ID } from '@angular/core';
+import { Inject, Injectable, Injector, PLATFORM_ID, Self } from '@angular/core';
 import { isNotNil } from '@ngxs-labs/data/internals';
 import { NEED_SYNC_TYPE_ACTION, NGXS_DATA_EXCEPTIONS } from '@ngxs-labs/data/tokens';
 import {
@@ -8,6 +8,7 @@ import {
     ExistingEngineProvider,
     PersistenceProvider,
     RootInternalStorageEngine,
+    StorageContainer,
     StorageMeta,
     UseClassEngineProvider
 } from '@ngxs-labs/data/typings';
@@ -26,47 +27,62 @@ import { PlainObject } from '@ngxs/store/internals';
 import { fromEvent } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
-/**
- * @privateApi
- */
-@Injectable()
-export class NgxsDataStorageEngine implements NgxsPlugin, RootInternalStorageEngine {
-    public static providers: Set<PersistenceProvider> = new Set();
-    private static keys: Map<string, void> = new Map();
+import { NGXS_DATA_STORAGE_CONTAINER_TOKEN } from './tokens/ngxs-data-storage-container';
 
-    constructor(@Inject(PLATFORM_ID) private _platformId: string, private injector: Injector) {
-        if (!isPlatformServer(this._platformId)) {
-            fromEvent(window, 'storage').subscribe((event: Any) => {
-                if (NgxsDataStorageEngine.keys.has((event as StorageEvent).key!)) {
-                    this.store!.dispatch({ type: NEED_SYNC_TYPE_ACTION });
-                }
-            });
-        }
+@Injectable()
+export class NgxsDataStoragePlugin implements NgxsPlugin, RootInternalStorageEngine {
+    public static injector: Injector | null = null;
+
+    constructor(@Inject(PLATFORM_ID) private _platformId: string, @Self() injector: Injector) {
+        NgxsDataStoragePlugin.injector = injector;
+        this.listenWindowEvents();
     }
 
     public get store(): Store | null {
-        return this.injector.get(Store, null);
+        return NgxsDataStoragePlugin.injector!.get(Store, null);
     }
 
     public get size(): number {
         return this.providers.size;
     }
 
+    public get container(): StorageContainer | never {
+        let container: StorageContainer;
+
+        try {
+            container = NgxsDataStoragePlugin.injector?.get(NGXS_DATA_STORAGE_CONTAINER_TOKEN)!;
+        } catch (e) {
+            throw new Error(NGXS_DATA_EXCEPTIONS.NGXS_PERSISTENCE_CONTAINER);
+        }
+
+        return container!;
+    }
+
     public get providers(): Set<PersistenceProvider> {
-        return NgxsDataStorageEngine.providers;
+        return this.container.providers;
+    }
+
+    public get keys(): Map<string, void> {
+        return this.container.keys;
     }
 
     public get entries(): IterableIterator<[PersistenceProvider, PersistenceProvider]> {
         return this.providers.entries();
     }
 
-    public static getProvidedKeys(): string[] {
-        return Array.from(NgxsDataStorageEngine.keys.keys());
-    }
+    private static exposeEngine(provider: PersistenceProvider): DataStorageEngine {
+        const engine: DataStorageEngine | null | undefined =
+            (provider as ExistingEngineProvider).existingEngine ||
+            NgxsDataStoragePlugin.injector!.get<DataStorageEngine>(
+                ((provider as UseClassEngineProvider).useClass as Any)!,
+                null!
+            );
 
-    public static clear(): void {
-        NgxsDataStorageEngine.keys.clear();
-        NgxsDataStorageEngine.providers.clear();
+        if (!engine) {
+            throw new Error(`${NGXS_DATA_EXCEPTIONS.NGXS_PERSISTENCE_ENGINE}:::${provider.path}`);
+        }
+
+        return engine;
     }
 
     public handle(states: PlainObject, action: ActionType, next: NgxsNextPluginFn): NgxsNextPluginFn {
@@ -81,18 +97,18 @@ export class NgxsDataStorageEngine implements NgxsPlugin, RootInternalStorageEng
 
         if (canBeSyncStoreWithStorage) {
             for (const [provider] of this.entries) {
-                const engine: DataStorageEngine = this.exposeEngine(provider);
+                const engine: DataStorageEngine = NgxsDataStoragePlugin.exposeEngine(provider);
                 const key: string = this.ensureKey(provider);
                 const value: string | undefined = engine.getItem(key);
                 if (isNotNil(value)) {
                     try {
                         const data: string | undefined | null = this.deserialize(value);
                         if (isNotNil(data) || provider.nullable) {
-                            NgxsDataStorageEngine.keys.set(key);
+                            this.keys.set(key);
                             states = setValue(states, provider.path, data);
                         } else {
                             engine.removeItem(key);
-                            NgxsDataStorageEngine.keys.delete(key);
+                            this.keys.delete(key);
                         }
                     } catch {
                         console.error(`${NGXS_DATA_EXCEPTIONS.NGXS_PERSISTENCE_DESERIALIZE}:::${provider.path}`);
@@ -107,13 +123,13 @@ export class NgxsDataStorageEngine implements NgxsPlugin, RootInternalStorageEng
                     const prevData: Any = getValue(states, provider.path);
                     const newData: Any = getValue(nextState, provider.path);
                     if (prevData !== newData || isInitAction) {
-                        const engine: DataStorageEngine = this.exposeEngine(provider);
+                        const engine: DataStorageEngine = NgxsDataStoragePlugin.exposeEngine(provider);
 
                         try {
                             const data: Any = this.serialize(newData, provider);
                             const key: string = this.ensureKey(provider);
                             engine.setItem(key, data);
-                            NgxsDataStorageEngine.keys.set(key);
+                            this.keys.set(key);
                         } catch (e) {
                             console.error(`${NGXS_DATA_EXCEPTIONS.NGXS_PERSISTENCE_SERIALIZE}:::${provider.path}`);
                         }
@@ -144,15 +160,13 @@ export class NgxsDataStorageEngine implements NgxsPlugin, RootInternalStorageEng
         return `${provider.prefixKey}${provider.path}`;
     }
 
-    private exposeEngine(provider: PersistenceProvider): DataStorageEngine {
-        const engine: DataStorageEngine | null | undefined =
-            (provider as ExistingEngineProvider).existingEngine ||
-            this.injector.get<DataStorageEngine>(((provider as UseClassEngineProvider).useClass as Any)!, null!);
-
-        if (!engine) {
-            throw new Error(`${NGXS_DATA_EXCEPTIONS.NGXS_PERSISTENCE_ENGINE}:::${provider.path}`);
+    private listenWindowEvents(): void {
+        if (!isPlatformServer(this._platformId)) {
+            fromEvent(window, 'storage').subscribe((event: Any) => {
+                if (this.keys.has((event as StorageEvent).key!)) {
+                    this.store!.dispatch({ type: NEED_SYNC_TYPE_ACTION });
+                }
+            });
         }
-
-        return engine;
     }
 }